+++ env.sh
@@ -1,2 +1,3 @@
-export YELP=ucb20Uq6G4KNJ7EHnndz-S6qwsx9g4aK3vfqGyKFn_b1j83D9-woDvbViMA1s47wt9yc4pDf7tG1KL_7UuFHZuo0Tiz3PJ1Ds39aI2236PZ6amiUxE_FaHXtkJ2WXXYx
+export YELP_KEY=ucb20Uq6G4KNJ7EHnndz-S6qwsx9g4aK3vfqGyKFn_b1j83D9-woDvbViMA1s47wt9yc4pDf7tG1KL_7UuFHZuo0Tiz3PJ1Ds39aI2236PZ6amiUxE_FaHXtkJ2WXXYx
 export EVENT_KEY=bgw8NQ28vRcNxKHB
+export WEATHER_KEY=9db10307657b0ff8224b0da642ac57f7
