+++ event_api.py
@@ -1,17 +1,6 @@
 import requests
 import os
-
-# Create event class to safe all the objects
-class EventData:
-    
-    def __init__(self, name, place, address, date):
-        self.name = name
-        self.place = place
-        self.address = address
-        self.date = date
-
-    def __str__(self):
-        return f'Name: {self.name} | Place: {self.place} | Address: {self.address} | Date: {self.date} \n'
+from event_class import * 
 
 def getEvents(location, keyword):
 
@@ -42,4 +31,4 @@ def getEvents(location, keyword):
         return event_list
 
     except Exception as e:
-        raise e
+        print('Please enter valid city name')
