+++ restaurant_api.py
@@ -1,20 +1,8 @@
 import requests
 import os
+from restaurant_class import *
 ##import yelp api key
-yelp_key = os.environ.get('YELP')
-
-class RestaurantData():
-    def __init__(self, name, location, restaurant_type, rating, price):
-        self.name = name
-        self.location = location
-        self.restaurant_type = restaurant_type
-        self.rating = rating
-        self.price = price
-    ##printing the restaurant information
-    def __str__(self):
-        return f'Name: {self.name} Location: {self.location} Type: {self.restaurant_type} Rating: {self.rating} Price: {self.price}\n'
-        
-
+yelp_key = os.environ.get('YELP_KEY')
 
 def getRestaurants(location):
     try:
@@ -43,4 +31,4 @@ def getRestaurants(location):
         ##returning the list of classes
         return restaurant_list
     except Exception as e:
-        raise e  
+        print('Please enter valid city name')
